move scripts out of subfolder
