+++ search_backend/services.py
@@ -3,9 +3,9 @@ from urllib.parse import urlparse
 from haystack_integrations.document_stores.opensearch import OpenSearchDocumentStore
 from opensearchpy import OpenSearch, Urllib3HttpConnection, Urllib3AWSV4SignerAuth
 
-from search_backend.api.lib.config import get_config
-from search_backend.api.lib.aws import get_aws_session
-from search_backend.api.lib.s3client import S3Client
+from search_backend.config import get_config
+from search_backend.aws import get_aws_session
+from search_backend.s3client import S3Client
 
 
 # S3 needs a specific region if we're using Analytical Platform buckets
