+++ tests/indexing_pipeline_test.py
